Moving i18n shared code into __init__.py instead.
