+++ kalite/i18n/management/commands/scrape_exercises.py
@@ -17,9 +17,9 @@ from optparse import make_option
 from django.core.management.base import BaseCommand, CommandError
 
 import settings
+from i18n import get_dubbed_video_map, lcode_to_ietf, lcode_to_django_lang, get_localized_exercise_dirpath
 from settings import LOG as logging
 from shared.topic_tools import get_node_cache
-from shared.i18n import get_dubbed_video_map, lcode_to_ietf, lcode_to_django_lang, get_localized_exercise_dirpath
 from utils.general import ensure_dir
 
 AVAILABLE_EXERCISE_LANGUAGE_CODES = ["da", "he", "pt-BR", "tr", "es", "fr"]
