+++ kalite/i18n/management/commands/update_po.py
@@ -21,8 +21,8 @@ from django.core.management import call_command
 from django.core.management.base import BaseCommand, CommandError
 
 import settings
+from i18n import lcode_to_django_dir, update_jsi18n_file
 from settings import LOG as logging
-from shared.i18n import lcode_to_django_dir, update_jsi18n_file
 from utils.django_utils import call_command_with_output
 from utils.general import ensure_dir
