+++ kalite/main/api_views.py
@@ -26,10 +26,10 @@ from . import topicdata
 from .api_forms import ExerciseLogForm, VideoLogForm, DateTimeForm
 from .models import VideoLog, ExerciseLog
 from config.models import Settings
+from i18n import lcode_to_ietf
 from securesync.models import FacilityGroup, FacilityUser
 from shared.caching import backend_cache_page
 from shared.decorators import allow_api_profiling, require_admin
-from shared.i18n import lcode_to_ietf
 from shared.jobs import force_job, job_status
 from shared.topic_tools import get_flat_topic_tree
 from shared.videos import delete_downloaded_files
