+++ kalite/main/tests/api_tests.py
@@ -7,11 +7,11 @@ from django.core.management.base import CommandError
 from django.core.urlresolvers import reverse
 from django.utils import unittest
 
+import i18n
 import settings
 from .base import MainTestCase
 from main.models import VideoLog, ExerciseLog
 from securesync.models import Facility, FacilityUser
-from shared import i18n
 from shared.testing import distributed_server_test, KALiteClient, KALiteTestCase
