+++ kalite/management/commands/zip_kalite.py
@@ -10,8 +10,8 @@ from django.core.management.base import BaseCommand, CommandError
 from django.core.management import call_command
 
 import settings
+from i18n import CROWDIN_CACHE_DIR, get_dubbed_video_map
 from securesync.models import Device
-from shared.i18n import CROWDIN_CACHE_DIR, get_dubbed_video_map
 from shared.khanload import KHANLOAD_CACHE_DIR
 from utils.general import ensure_dir
 from utils.platforms import is_windows, not_system_specific_scripts, system_specific_zipping, _default_callback_zip
