+++ kalite/shared/caching.py
@@ -16,9 +16,10 @@ from django.views.decorators.cache import cache_control
 from django.views.decorators.cache import cache_page
 from django.views.decorators.http import condition
 
+import i18n
 import settings
 from settings import LOG as logging
-from shared import i18n, topic_tools
+from shared import topic_tools
 from updates.models import VideoFile
 from utils.internet import generate_all_paths
