+++ kalite/shared/topic_tools.py
@@ -7,9 +7,10 @@ from functools import partial
 
 from django.utils.translation import ugettext as _
 
+import i18n
 import settings
 from settings import LOG as logging
-from shared import i18n, khanload
+from shared import khanload
 from utils.general import softload_json
 
 CACHE_VARS = []
