+++ kalite/updates/management/commands/languagepackdownload.py
@@ -15,11 +15,11 @@ from django.utils.translation import ugettext as _
 import settings
 import version
 from .classes import UpdatesStaticCommand
+from i18n import LOCALE_ROOT, DUBBED_VIDEOS_MAPPING_FILEPATH
+from i18n import get_language_pack_metadata_filepath, get_language_pack_filepath, get_language_pack_url, get_localized_exercise_dirpath, get_srt_path
+from i18n import lcode_to_django_dir, lcode_to_ietf, update_jsi18n_file
 from settings import LOG as logging
 from shared import caching
-from shared.i18n import LOCALE_ROOT, DUBBED_VIDEOS_MAPPING_FILEPATH
-from shared.i18n import get_language_pack_metadata_filepath, get_language_pack_filepath, get_language_pack_url, get_localized_exercise_dirpath, get_srt_path
-from shared.i18n import lcode_to_django_dir, lcode_to_ietf, update_jsi18n_file
 from shared.videos import REMOTE_VIDEO_SIZE_FILEPATH
 from utils.general import ensure_dir
 from utils.internet import callback_percent_proxy, download_file
