+++ kalite/updates/views.py
@@ -25,11 +25,11 @@ import version
 from .models import VideoFile
 from config.models import Settings
 from control_panel.views import local_device_context, user_management_context
+from i18n import lcode_to_ietf, get_installed_language_packs, lang_best_name
 from securesync.models import Facility, FacilityUser, FacilityGroup, Device
 from securesync.views import require_admin, facility_required
 from shared import topic_tools
 from shared.decorators import require_admin
-from shared.i18n import lcode_to_ietf, get_installed_language_packs, lang_best_name
 from shared.jobs import force_job
 from utils.internet import am_i_online, JsonResponse
